+++ .php-cs-fixer.dist.php
@@ -3,7 +3,15 @@
 use PhpCsFixer\Runner\Parallel\ParallelConfigFactory;
 
 $finder = (new PhpCsFixer\Finder())
-    ->in([__DIR__.'/src', __DIR__.'/tests']);
+    ->in(__DIR__)
+    ->exclude('var')
+    ->notPath([
+        'config/bundles.php',
+        'config/preload.php',
+        'config/reference.php',
+        'public/index.php',
+    ])
+;
 
 return (new PhpCsFixer\Config())
     ->setUnsupportedPhpVersionAllowed(true)
@@ -18,9 +26,9 @@ return (new PhpCsFixer\Config())
         'protected_to_private' => true,
         'single_quote' => true,
         'header_comment' => [
-            'comment_type' => 'PHPDoc',
+            'comment_type' => 'comment',
             'location' => 'after_declare_strict',
-            'header' => "@copyright Copyright (c) 2020 - 2026 Communitales GmbH (https://www.communitales.com/)\n\nFor the full copyright and license information, please view the LICENSE\nfile that was distributed with this source code.",
+            'header' => "SPDX-FileCopyrightText: 2020 Communitales GmbH\n\nSPDX-License-Identifier: MIT",
         ],
 
     ])