+++ rector.php
@@ -2,6 +2,12 @@
 
 declare(strict_types=1);
 
+/*
+ * SPDX-FileCopyrightText: 2020 Communitales GmbH
+ *
+ * SPDX-License-Identifier: MIT
+ */
+
 use Rector\Config\RectorConfig;
 
 return RectorConfig::configure()
@@ -9,19 +15,14 @@ return RectorConfig::configure()
         __DIR__.'/src',
         __DIR__.'/tests',
     ])
-    ->withSkip([
-        __DIR__.'/src/Entity',   // SetList::CODE_QUALITY macht da Mist
-    ])
     ->withPreparedSets(
         deadCode: true,
         codeQuality: true,
         codingStyle: true,
         typeDeclarations: true,
         phpunitCodeQuality: true,
-        doctrineCodeQuality: true,
-        symfonyCodeQuality: true
     )
     ->withPhpSets(php84: true)
-    ->withComposerBased(twig: true, doctrine: true, phpunit: true, symfony: true)
+    ->withComposerBased(phpunit: true)
     ->withImportNames()
     ->withCache('var/cache/rector');