+++ src/StatusBus/StatusLevel.php
@@ -10,7 +10,10 @@ declare(strict_types=1);
 
 namespace Communitales\Component\StatusBus;
 
-interface StatusBusInterface
+enum StatusLevel: string
 {
-    public function publish(StatusMessage $message): void;
+    case Error = 'error';
+    case Info = 'info';
+    case Success = 'success';
+    case Warning = 'warning';
 }